requirements.txt
pyodbc==4.0.30
configparser==5.0.0
flask==1.1.2
python-dateutil==2.8.1
jsonschema==3.2.0
pysmb==1.2.4
pdfdocument==4.0.0
simplejson==3.17.2